requirements.txt
requests
python-dotenv
SQLAlchemy
psycopg2-binary